changes for two-stage hypershift IC upgrade
The major difference between managed and self hosted IC upgrade is the
stateful set on the master side. patch the update so the commit works
for this too.

HACK to get the stateful set updated was to remove it when safe to do so
in order to allow the CNO to create a version updated as we need to in
phase 2